backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
redis>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0

